Fonctionnement parfait entre deux gros champions

+++ includes/vm.h
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/10 12:13:30 by nsalle            #+#    #+#             */
-/*   Updated: 2020/05/09 11:45:06 by user42           ###   ########lyon.fr   */
+/*   Updated: 2020/05/09 13:28:19 by user42           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ typedef struct      	s_proclist
 	uint8_t				carry;
 	uint8_t				alive;
 	uint8_t				param[4];
-	int					last_alive;
+	long long			last_alive;
 	int					reg[REG_NUMBER];
 }                   	t_proclist;
 
@@ -74,7 +74,7 @@ typedef struct      	s_vm
 	t_proclist			*beginactive;
 	uint32_t			pid_tracker;
 	t_player			players[MAX_PLAYERS];
-	uint64_t			cycles;
+	long long			cycles;
 	uint8_t				game;
 }                   	t_vm;
